+++ src/main/resources/index.d.ts
@@ -6,7 +6,6 @@ export {
 
 // Global-modifying module, should be placed in moduleRoot/index.d.ts
 /*declare global {
-	// We don't want any global objects, pass around JavaBridge instead
-	//const app :App;
-	//const log :Log;
+	//const app: App;
+	//const log: Log;
 }*/
