--- src/main/resources/lib/explorer/_uncoupled/document/index.ts
@@ -1,19 +1,0 @@
-export {addExtraFieldsToDocumentType} from './addExtraFieldsToDocumentType';
-export {applyDocumentTypeToDocumentNode} from './applyDocumentTypeToDocumentNode';
-export {cleanData} from './cleanData';
-export {constrainPropertyNames} from './constrainPropertyNames';
-export {create} from './create';
-export {createOrUpdate} from './createOrUpdate';
-export {
-	addMissingSetToFieldsArray,
-	applyDefaultsToField,
-	fieldsArrayToObj,
-	fieldsObjToArray,
-	isField,
-	isFields
-} from './field';
-export {typeCastToJava} from './typeCastToJava';
-export {update} from './update';
-export {validate} from './validate';
-export {validateOccurrences} from './validateOccurrences';
-export {validateTypes} from './validateTypes';
