--- src/main/resources/lib/explorer/_uncoupled/documentType/documentTypeNameToPath.ts
@@ -1,3 +1,0 @@
-import {PATH_DOCUMENT_TYPE_FOLDER} from '/lib/explorer/documentType/constants';
-
-export const documentTypeNameToPath = (documentTypeName: string) => `${PATH_DOCUMENT_TYPE_FOLDER}/${documentTypeName}`;
