--- src/main/resources/lib/explorer/_uncoupled/documentType/index.ts
@@ -1,1 +1,0 @@
-export {update} from './update';
