--- src/main/resources/lib/explorer/_uncoupled/index.ts
@@ -1,4 +1,0 @@
-export * from '../constants';
-export * as document from './document/index';
-export * as documentType from './documentType/index';
-export * as repo from './repo/index';
