--- src/main/resources/lib/explorer/_uncoupled/repo/index.ts
@@ -1,1 +1,0 @@
-export {list} from './list';
