+++ src/main/resources/lib/explorer/documentType/documentTypeNameToPath.ts
@@ -1,1 +1,3 @@
-export {documentTypeNameToPath} from '/lib/explorer/_uncoupled/documentType/documentTypeNameToPath';
+import {PATH_DOCUMENT_TYPE_FOLDER} from '/lib/explorer/documentType/constants';
+
+export const documentTypeNameToPath = (documentTypeName: string) => `${PATH_DOCUMENT_TYPE_FOLDER}/${documentTypeName}`;
