+++ src/main/resources/lib/explorer/documentType/getDocumentTypeByName.ts
@@ -1,5 +1,13 @@
-import {getDocumentTypeByName as pureTsGetDocumentTypeByName} from '/lib/explorer/_uncoupled/documentType/getDocumentTypeByName';
-import {javaBridge} from '../_coupling/javaBridge';
+import type {DocumentTypeNode} from '/lib/explorer/types/index.d';
+
+
+import {connect} from '/lib/xp/node';
+import {
+	PRINCIPAL_EXPLORER_READ,
+	REPO_ID_EXPLORER
+} from '../constants';
+import {coerseDocumentType} from './coerseDocumentType';
+import {documentTypeNameToPath} from './documentTypeNameToPath';
 
 
 export function getDocumentTypeByName({
@@ -7,5 +15,15 @@ export function getDocumentTypeByName({
 }: {
 	documentTypeName: string
 }) {
-	return pureTsGetDocumentTypeByName({documentTypeName}, javaBridge);
+	const explorerReadConnection = connect({
+		branch: 'master',
+		principals: [PRINCIPAL_EXPLORER_READ],
+		repoId: REPO_ID_EXPLORER
+	});
+	const path = documentTypeNameToPath(documentTypeName);
+	const documentTypeNode = explorerReadConnection.get(path) as DocumentTypeNode;
+	if (!documentTypeNode) {
+		throw new Error(`Unable to get documentType with name:${documentTypeName}!`);
+	}
+	return coerseDocumentType(documentTypeNode);
 }
